pyproject.toml
[project]
name = "bouncer"
version = "0.1.0"
description = "Biological data quality contract layer for Nextflow pipeline outputs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # CLI + terminal
    "typer>=0.12.0",
    "rich>=13.7.0",
    # HTTP client (CLI → Modal API)
    "httpx>=0.27.0",
    # Config + validation
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "packaging>=24.0",
]

[project.optional-dependencies]
# Heavy deps only needed inside Modal containers — not installed locally
container = [
    "anthropic>=0.25.0",
    "anndata>=0.10.6",
    "scanpy>=1.10.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pandera>=0.19.0",
    "duckdb>=0.10.0",
    "pdfplumber>=0.10.3",
    "fcsparser>=0.2.8",
    "fastapi>=0.111.0",
    "python-multipart>=0.0.9",
]
infra = [
    "modal>=0.62.0",
]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.100.0",
    "pytest-asyncio>=0.23.0",
    "modal>=0.62.0",
]

[project.scripts]
bouncer = "bouncer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "hypothesis>=6.100.0",
    "pytest-asyncio>=0.23.0",
    "modal>=0.62.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
